refactor: connecting ES-modules to project
